refactor: move GraphQL operations into a separate folder
