+++ imports/plugins/core/dashboard/client/graphql/mutations/updateShop.js
@@ -1,0 +1,13 @@
+import gql from "graphql-tag";
+import shopCommon from "../fragments/shopCommon";
+
+export default gql`
+  mutation updateShop($input: UpdateShopInput!) {
+    updateShop(input: $input) {
+      shop {
+        ...ShopCommon
+      }
+    }
+  }
+  ${shopCommon}
+`;
