+++ imports/plugins/core/dashboard/client/graphql/queries/shop.js
@@ -1,0 +1,11 @@
+import gql from "graphql-tag";
+import shopCommonFragment from "../fragments/shopCommon";
+
+export default gql`
+  query shopQuery($id: ID!) {
+    shop(id: $id) {
+      ...ShopCommon
+    }
+  }
+  ${shopCommonFragment}
+`;
